requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
jupyterlab
joblib
kafka-python
pyyaml
pdfminer.six>=20201018
pdfminer.six
camelot-py[cv]
pandas
openpyxl
pip install -r requirements.txt
